+++ src/protocols/valve/mod.rs
@@ -1,4 +1,7 @@
+
+/// The implementation.
 pub mod protocol;
+/// All types used by the implementation.
 pub mod types;
 
 pub use protocol::*;